🔧 chore(.gitignore): add jcacperalta.github.io-master/ and sam_learner_projects-master/ to .gitignore 📝 docs(work.html): update job title and add new responsibility in work experience section
The .gitignore file has been updated to ignore the directories jcacperalta.github.io-master/ and sam_learner_projects-master/. This is done to prevent these directories from being tracked by git, as they contain project files that are not relevant to the current repository.

In the work.html file, the job title has been updated from 'Full-stack Developer' to 'Developer 2' to reflect the recent promotion. A new responsibility has been added to the job description to provide a more accurate representation of the role.