🔧 refactor(app.vue): replace NuxtWelcome with NuxtLayout and NuxtPage for better structure 🔥 remove(nuxt.config.ts): remove redundant nuxt.config.ts file ✨ feat(nuxt.config.js): add new nuxt.config.js file with configuration for devtools, srcDir, css, and image 🎨 style(main.css): add new main.css file with styles for root, header, main, nav links, body, footer, responsive media, and animation ✨ feat(Generic.vue): add new Generic.vue layout with PersonalHeader component ✨ feat(index.vue): add new index.vue page with SEO meta and page meta
The changes were made to improve the structure of the application by replacing the NuxtWelcome component with NuxtLayout and NuxtPage in app.vue. The nuxt.config.ts file was removed as it was redundant and a new nuxt.config.js file was added with configuration for devtools, srcDir, css, and image. A new main.css file was added with styles for various elements and responsive media. A new Generic.vue layout was added with a PersonalHeader component for better layout management. Lastly, a new index.vue page was added with SEO meta and page meta for better SEO optimization.