🔧 refactor(nuxt.config.js): add 'app' configuration for better SEO 🔥 remove(public/favicon.ico, src/assets/images/jclogofix-smaller.svg): delete unused favicon and logo files ✨ feat(src/components/NavLinks.vue): create new NavLinks component for better navigation 🔧 refactor(src/components/PersonalHeader.vue): replace 'a' tags with 'NuxtLink' for better SPA navigation 🔧 refactor(src/pages/index.vue): add NavLinks component to the main page for better user experience 🎨 style(src/public/*): add new favicon and logo files for better branding
The 'app' configuration was added to the nuxt.config.js file to improve SEO by specifying the language and adding a link to the stylesheet. Unused favicon and logo files were removed to clean up the project. A new NavLinks component was created to improve navigation within the site. 'a' tags in the PersonalHeader.vue file were replaced with 'NuxtLink' to improve navigation in the SPA. The NavLinks component was added to the main page to improve user experience. New favicon and logo files were added to improve the branding of the site.